+++ src/ingest_validation_tools/plugin_validator.py
@@ -94,11 +94,11 @@ def run_plugin_validators_iter(metadata_path: PathOrStr,
 def validation_error_iter(base_dir: PathOrStr, assay_type: str,
                           plugin_dir: PathOrStr) -> Iterator[KeyValuePair]:
     """
-    Given a base directory pointing to a tree of submission data files and a
+    Given a base directory pointing to a tree of upload data files and a
     path to a directory of Validator plugins, iterate over the results of
     applying all the plugin validators to the directory tree.
 
-    base_dir: the root of a directory tree of submission data files
+    base_dir: the root of a directory tree of upload data files
     assay_type: the assay type which produced the data in the directory tree
     plugin_dir: path to a directory containing classes derived from Validator